+++ django_file_form/widgets/__init__.py
@@ -1,0 +1,2 @@
+from .upload_widget import UploadWidget
+from .upload_multiple_widget import UploadMultipleWidget
